+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.0 branch
 ### 1.0.2
+* added `get_class_short_name()` global function;
 * fixed `rtr()` global function. It can also use the `ROOT` environment variable.
 
 ### 1.0.1